+++ 1_linked_list/2_doubly_linked_list/list_remove.c
@@ -4,7 +4,7 @@ int	list_remove(t_linked_list *list, int n)
 {
 	t_node	*curr;
 
-	if (list == 0 || n < 0 || n >= list->size || (curr = list_get(list, n)) == 0)
+	if (list == 0 || n < 0 || n >= (int)list->size || (curr = list_get(list, n)) == 0)
 		return (0);
 	if (curr->prev == 0 || n == 0)
 	{
@@ -12,7 +12,7 @@ int	list_remove(t_linked_list *list, int n)
 		if (curr->next)
 			(curr->next)->prev = 0;
 	}
-	if (curr->next == 0 || n == list->size)
+	if (curr->next == 0 || n == (int)list->size)
 	{
 		list->tail = curr->prev;
 		if (curr->prev)
